In frequence conversion the scaling factor of float type was determined by dividing two integers. That couldn't play out well.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@16847 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/sound/loader/common/soundraw.cpp
@@ -162,8 +162,9 @@ void *ConvertChannels(void *d, const csSoundFormat *oldfmt,
 void *ConvertFreq(void *d, const csSoundFormat *oldfmt,
   const csSoundFormat *newfmt, long &NumSamples)
   {
-  float Factor=newfmt->Freq/oldfmt->Freq;
+  float Factor=newfmt->Freq/(float)oldfmt->Freq;
   unsigned long NewNumSamples=(unsigned long)(NumSamples*Factor);
+
   if (oldfmt->Bits==16)
   {
     CONVERT_FREQ_TYPE(short,oldfmt->Channels);
